pyproject.toml
[project]
name = "agentic-hello-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "ipykernel>=7.2.0",
    "openai-agents>=0.13.6",
    "psycopg[binary]>=3.3.3",
    "sqlmodel>=0.0.38",
]
